requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
asyncpg==0.30.0
black==24.10.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
flake8==7.1.1
Flake8-pyproject==1.2.3
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pg8000==1.31.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
scramp==1.4.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
